Inserted the functions of copying resources from name-generator-unit-tests and nickname-generator-unit-tests into the if to avoid errors
Added the if conditional to name-generator-unit-tests and nickname-generator-unit-tests because before there was an error when trying to delete resources folders again

Changed csgocore.cpp/new_game() return from void to bool and added a try/catch

Added assert::istrue  to initialize() and new_game() in unit tests

Fixed messages in csgo-core-unit-tests

+++ tests/nickname-generator-unit-tests.cpp
@@ -30,11 +30,11 @@ TEST_CLASS (nicknamegenerator)
                 Assert::IsTrue(std::filesystem::create_directory(_current_path /
                                                                  "resources"),
                                L"Failed to create temp directory");
-            }
 
-            std::filesystem::copy(nickname_resource_dir,
-                                  _current_path / R"(resources/nickname)",
-                                  std::filesystem::copy_options::recursive);
+                std::filesystem::copy(nickname_resource_dir,
+                                      _current_path / R"(resources/nickname)",
+                                      std::filesystem::copy_options::recursive);
+            }
 
             Assert::IsTrue(
                 std::filesystem::exists(_current_path /
@@ -234,15 +234,25 @@ TEST_CLASS (nicknamegenerator)
 
         try
         {
-            Assert::IsTrue(std::filesystem::remove_all(
-                               _current_path / "resources/nickname") > 0,
-                           L"Failed to remove resources");
-            Assert::IsFalse(std::filesystem::exists(_current_path /
-                                                    R"(resources/nickname)"),
-                            L"Nickname folder still exists");
+            if (std::filesystem::exists(_current_path / "resources/nickname"))
+            {
 
-            Logger::WriteMessage("Success cleaning up the resources folder for "
-                                 "nickname generator tests");
+                Assert::IsTrue(std::filesystem::remove_all(
+                                   _current_path / "resources/nickname") > 0,
+                               L"Failed to remove resources");
+                Assert::IsFalse(std::filesystem::exists(
+                                    _current_path / R"(resources/nickname)"),
+                                L"Nickname folder still exists");
+
+                Logger::WriteMessage(
+                    "Success cleaning up the resources folder for "
+                    "nickname generator tests");
+            }
+            else
+            {
+                Logger::WriteMessage(
+                    "Resources already deleted by csgo-core-unit-tests");
+            }
         }
         catch (const std::exception&)
         {
